+++ src/covid/Statistic.java
@@ -15,24 +15,12 @@ public class Statistic {
         return countryName;
     }
 
-    public void setCountryName(String name){
-        countryName = name;
-    }
-
     public String getCountryCode(){
         return countryCode;
     }
 
-    public void setCountryCode(String code){
-        countryCode = code;
-    }
-
     public String getCountryContinent(){
         return countryContinent;
-    }
-
-    public void setCountryContinent(String continent){
-        countryContinent = continent;
     }
 
 }
